+++ packages/system-functions/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stoker-platform/system-functions
 
+## 0.5.84
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.51
+    - @stoker-platform/node-client@0.5.72
+    - @stoker-platform/utils@0.5.63
+
 ## 0.5.83
 
 ### Patch Changes